Bug Fixes
Line 116:
allowMismatchBlank would set the value to blank, but the flexselect
would still reset the displayed selection if allowMismatch was false.
Add else to second if statement so both do not occur

Lines 119-120:
Remove dropdownList.show() on blur of flexselect when
hideDropdownOnEmptyInput=true, seems to serve no purpose.

Lines 239-245:
Move hide/show logic of hideDropdownOnEmptyInput to the end of
filterResults() instead of keyUp(). This will now apply when a
flexselect is clicked active.
Change dropdownList to dropdown, so an empty box with the border still
visible does not show